[clean]得点計算関数を外部に切り出し
